+++ CHANGELOG.md
@@ -10,6 +10,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - Converted `<manifold-data-provision-button>` to use GraphQL data (#600)
+- Converted `<manifold-data-rename-button>` to use GraphQL (#596)
 
 ### Breaking Changes: